backend/pyproject.toml
[tool.poetry]
name = "picasso test"
version = "1.0"
description = "Picasso Test"
authors = ["Александр Рассыхаев <farexzon@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = ">=3.2,<4.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
pre-commit = "^3.4.0"
celery = "^5.3.4"
psycopg2-binary = "^2.9.7"
redis = "^5.0.0"
pillow = "^10.0.1"
pymupdf = "^1.23.3"
pytest = "^7.4.2"
pytest-django = "^4.5.2"


[tool.poetry.group.test.dependencies]
pre-commit = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
